Adds instructor section (which is wide for now maybe permanently?) and radial pie chart for moods. Adds moods API and parser to get data in D3 format. Installs D3.
