+++ client/src/instructor/Results.js
@@ -1,0 +1,12 @@
+import MoodChart from './MoodChart'
+
+const Results = () => {
+  return (
+    <div>
+      List of Charts
+      <MoodChart />
+    </div>
+  );
+}
+
+export default Results
